Issue #2425: WIP
While dirty hack works for the whole deserialisation,
it ouputs a lot of zeros at the end, I am not sure if that would
worke well in the future, so its better to make sure everything works as
expected. I have a few examples using different array sizes which can't
be calculated between them with the same formula, I think I need to
include an extra class from m2g-rs232 to handle it (additional
abstraction).